backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
statsmodels
prophet
python-multipart
python-dotenv
pydantic
openai
